refactor: divide sidebar component from Semester Page to separate file
